+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.191](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.190...@synapsecns/synapse-interface@0.1.191) (2023-11-29)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.190](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.189...@synapsecns/synapse-interface@0.1.190) (2023-11-28)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.189](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.188...@synapsecns/synapse-interface@0.1.189) (2023-11-22)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.188](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.187...@synapsecns/synapse-interface@0.1.188) (2023-11-22)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.1.187](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.186...@synapsecns/synapse-interface@0.1.187) (2023-11-21)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface